Fix Bug 2101 - unable to fetch/send mail from chandler with the latest build
The new notification code was holding a reference to a ref-collection
(XMLRefDict) which was causing threading problems with Mail.  The new
code stores UUID and attribute names and dereferences the
ref-collection prior to query evaluation.

Another threading related problem was that callbacks were being
processed without regard to the view containing the query.

Expanded unit tests to cover related cases


git-svn-id: http://svn.osafoundation.org/chandler/trunk@3680 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe