Schema API: added script that reports on current porting status of all parcels.  To run it, type:
    RunPython -m schema_status

And review the output.  This is basically an automated "to-do"
list for porting parcels to the schema API.  When porting parcels,
run this to verify that your changes actually took something off
of one of the problem lists, before you run all the tests and
find out you forgot something.  :)

git-svn-id: http://svn.osafoundation.org/chandler/trunk@5626 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe