- First cut at readonlyness in the detail view; see notes below. Fixes bug 2976 (Calendar events from read-only shares should be read-only in the detail view) and bug 2974 (Don't show empty fields in read-only events). - Introduced a typed-property mechanism, currently named Computed, that I'm hoping will get merged into schema (additional note below).
Detail view changes:
- Some areas (the horizontal ContentItemDetail block slices that contain a label and AEBlock) will hide themselves if 
the attribute they contain isn't modifiable and has no value: location, alarm, and the still-hidden recurrence fields 
work this way. The all-day checkbox is also hidden if it's not checked, per the spec. Title and the note body always 
show themselves, even if they have no value - things look too weird when they're hidden).
- The "to" label and the event start & end _time_ fields are hidden for readonly allday or anytime events
- The stamping and 'private' markupbar buttons are disabled if the itsKind/isPrivate attributes aren't modifiable, 
respectively. Unfortunately, the toolbar doesn't do a good job of showing this disabledness (on PC, anyway)

Attribute editor changes:
- selection mechanism now includes readonlyness, which allows us to produce a different AE when an item's attribute is 
readonly (for instance: we now use a dropdown menu for event status, which is disabled when the item is readonly. If 
we'd prefer to display that as a static string, we can tweak it in XML to use a different AE instead of having to 
customize the AE itself.)
- the AEs that display using a wx.TextCtrl specify the wx flag TE_READONLY when they're not modifiable. Unfortunately, 
the native widgets act differently on different platforms when this flag is set (on Mac, it appears that only rich text 
fields can have text selected when readonly - the other textctrls don't seem to listen to selection events.)

Computed:
- We used the python property() mechanism for some computed values on items (eg, duration and reminderDelta); however, 
it was problematic that these properties weren't typed like schema attributes are. With PJE's help, I added a Computed() 
mechanism that subclasses property() to add a schema type (as well as displayName) - this lets them be mapped to 
attribute editors more cleanly.
- For now, Computed() is declared in ContentItem.py, since it's really mostly useful for letting computed attributes on 
ContentItems be exposed by attribute editors. I'm hoping that PJE will choose to integrate it into schema.py, in which 
case it'd move.



git-svn-id: http://svn.osafoundation.org/chandler/trunk@6101 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe