Fix 3743: Convert detail view parcel.xml to python.
Note: I didn't convert the item-collection detail view, since it's not currently used (so there's no way to test it!); when (if) we want to display an item collection detail view again, we can convert it.

Also note: This is essentially a raw conversion. I'm going to write a few utility functions that parcel authors can use - to ensure spacing consistency, etc.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@6654 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe