Fix for bug 3403 (Create a default item on the calendar).
The "Welcome" note is now an event which appears at Noon on the day you
first run Chandler.  It also no longer a photo for 3 reasons:  not everyone
is in that picture, there's not really enough room for an image when you
throw in the calendar event fields, and the scrollable image attribute editor
is buggy.



git-svn-id: http://svn.osafoundation.org/chandler/trunk@7064 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe