Changed the way version information is generated and stored in version.py The version number used to be stored in __hardhat__.py and written to version.py right before the distribution is created.  This was required because the install scripts needed to know the major, minor and release infomation.
Now the version information is read from version.py, parsed and passed on to the
various scripts.  Also now the SVN revision number is written to version.py.

The revision information is now needed by the performance output code so that the
build and revision data can be stored along with the performance numbers.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@7149 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe