Fix bug 4498 - schema.Annotation currently doesn't work correctly for core-schema kinds like "Kind".  Reviewed by John.
This change also removes the log message for 'syncing attribute
descriptors', which now causes a lot of output because annotating
'Kind' (which John needs to do), requires all of the core kind
classes to resync descriptors.  On IRC, Andi recommended removing
the message altogether, rather than lowering the priority of the
message.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@8130 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe