Fix for bug 4551 (deleting from not-in-my-calendar share puts the item in all) r=sprout
The problem was that adding an item to a not-mine collection's exclusions list
would instantly move that item into the All collection *before* I had a chance
to see if the item was in any other collections.  Now that check is done
earlier.

Also added a unit test for this.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@8176 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe