bug 2765, add progress gauge to import dialog bug 4387, default import directory to Desktop, and persist last import directory bug 4536, removed import time
r=morgen

Added progress gauge to the iCalendar import dialog.

Refactored showFileDialog out of ImportExport into dialog.Util, which seems like a more appropriate place for it.  In the process, fiddled with the import AccountPreferences line there,
because it was creating circular imports by importing sharing, moved the import into the mailError function

git-svn-id: http://svn.osafoundation.org/chandler/trunk@8294 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe