Fix for bug 4721, cutting fails, r=alecf
The first problem was an argument mismatch, quick to fix.  The larger problem was that for 0.6, cut of a recurring event can only really apply to the whole recurrence series.  Updated DeleteSelection and removeFromCollection to accept a cutting argument so that RecurrenceDialog could treat a delete differently from a cut.

git-svn-id: http://svn.osafoundation.org/chandler/trunk@8376 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe