Draft release of the Parcel Developer's Guide to the Schema API. This does *not* contain a complete API reference yet; this is just the "guide" part explaining all the concepts and how to use them. If you want to read this in HTML form, you can see preview versions at http://peak.telecommunity.com/DevCenter/ChandlerSchemaAPI which I periodically update with intermediate drafts to see how they look.  (Although the final output for the docs/ directory will use a very different stylesheet.)
This document is a doctest, which I run on my machine to verify
its correctness, but I am not yet checking in the changes to the
test suite that cause it to be run.  That will happen later, along
with some changes I've been making to the existing schema_api.txt
doctest.  So, for now, this doctest won't be run by the tinderboxes.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@8508 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe