Fixed issue in QALogger.py: Added self.File.flush() call to Print method in Testlogger.
Problem Description:
QALogger is a logger for CATS written from scratch, since it just creates file object and starts writing to it, it fills up a default buffer before actually writing to the file.
This caused problems when tests failed because the log stops before the test ran.
To fix this I just added flush to the Print method inside the TestLogger class.

git-svn-id: http://svn.osafoundation.org/chandler/trunk@9113 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe