Bug 5399, if --profileDir is specified create the directory if it does not exist.
Also, if we encountered a startup problem that prevented us
from creating the application object Chandler just died silently;
now we create a PySimpleApp in those cases so that we can display
an error dialog.

Fixed two cases where the use of logging.exception was causing
us to report the traceback twice (fix was to change them to
logging.error calls).

r=robin.

git-svn-id: http://svn.osafoundation.org/chandler/trunk@9953 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe