Refactored ModifyCollectionEvent and NewCollectionEvent to become AddToSidebarEvent.
Upated flickr, evdb, certstore and new collection command to use the AddToSidebarEvent.

Added a preferredKind to collections which is used as a hint to specify where to display the collection.

When new collections are added to the sidebar the first item in the collection is selected and displayed in the detail view.

git-svn-id: http://svn.osafoundation.org/chandler/trunk@10208 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe