Fix bug 5697 (Changes to the timezonelist should update the timezone drop down), r=jeffrey.
- The new centralized item-watching mechanism now supports collection and item
notifications on the same block; it will also notify the widget and the block
if both have implementations of the handler method.
- To improve performance, the new mechanism now ignores notifications
for items in the process of being stamped (see comment in src).
- CalendarControl now uses the new mechanism instead of creating its
own item watchers.
- The timezone attribute editor now always rebuilds the menu when
setting the value, even if it hasn't changed. I'm concerned about this causing
a performance problem, since that method is called unnecessarily as it is;
however, I don't have a way to tell whether rebuilding the menu is actually
unnecessary.

git-svn-id: http://svn.osafoundation.org/chandler/trunk@10542 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe