Fix bug #6276: "Annotations can't point to themselves".
Annotations can point to themselves, they just couldn't do it using strings as
forward references.  This change makes it work for strings too.

Note that using strings to name types is an undocumented feature that was used
to assist porting the schema from XML.  Please see the schema API documentation
at:

http://chandler.osafoundation.org/docs/0.7/parcel-schema-guide.html#bi-directional-references

for an explanation of how to create bidirectional references without using        
strings to name types.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@11204 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe