Add --catch=(normal|tests|never) command-line option, which will shortly replace --nocatch...
"normal" is the default: use a top-level exception handler, as well as an exception handler in the functional-test loop that prevents one test failure from killing the whole run. 

"tests" is what --nocatch used to be: don't use the top-level exception handler (so the feedback alert won't appear), but do use the functional-test loop handler. 

"never" says use neither: it's what you want if you're trying to debug a functional test failure -- Wing will see a such an exception as unhandled, and stop (which is good!).

Also, change all the places where --nocatch was used, to the new equivalent --catch=tests)

git-svn-id: http://svn.osafoundation.org/chandler/trunk@11661 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe