Fix for bug 7041 (Intermittent tinderbox sharing test failure: 'NoneType' has no attribute 'itsView') r=grant
The tests were not properly handling any kind of publishing failure, instead
continuing on as if nothing happened.  This patch will now report an error in
that case.  Also, with many clients simultaneously running the tests, there
was an increasing chance of trying to create a collection of a given name at
the same time as someone else.  The tests have been changed to use a UUID in
the collection name to prevent this.

In alpha5 we'll take this a step further by retrying other collection names
until we find one that doesn't exist.  We already do something like this when
first coming up with a unique name, but if another client swoops in and creates
the collection right before we do, we simply raise an exception.  It was this
exception that wasn't being handled by the test framework.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@12003 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe