Fix bug 6996 (Collection indexes can be created on wrong attributes), r=john, by adding IndexDefinition items that can be looked up by collections when indexes are requested. (Previously, indexes were explicitly created by the UI, but if an index was needed at other times, it was possible that an index would be created on the wrong attributes.)
Index names are now also scoped by parcel name (and are no longer just an attribute name), so we won't have to worry about conflicts, which is nice.

This also removes the workaround for bug 6639, which disabled sorting on the stamping columns; however, the indexes defined here for these columns aren't the ones specified by the dashboard spec, which will be added in alpha5.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@12006 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe