- Finish bug 7251 on the trunk, use cosmo-exclude property for freebusy   exclusion, r=morgen
As it happens, Cosmo 0.5.0's freebusy rollup appears to be broken
(that's bug 7309), so the only difference this will make for now is that
when restoring collections, the mine/not-mine bit will be controlled
by the status of this property.  If users change mine-ness, sync will update
this on the server if the collection is shared by that user.
  


git-svn-id: http://svn.osafoundation.org/chandler/trunk@12264 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe