Fix for bug 7292 and bug 7331 - both concerning smaple text in the detail view 7292 was "Text widgets should all have localizable sampleText attributes" 7331 was "Add hint text to addressing fields"
The code said that an empty sample text would cause an item's displayName to be used,
but instead the text "displayName" was being used. This checkin does not address that
behaviour - perhaps it should be looked into? Isn't displayName going away?

Since 'a meaningful default' was requested by bug 7292, I've supplied "enter title"


git-svn-id: http://svn.osafoundation.org/chandler/trunk@12293 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe