Set up EmailAddressAttributeEditor to be a staticControlDelegate; renamed 'delegate' to 'staticControlDelegate'
This check in also should not change anything. It is setting the stage for the indicator checkin,
which will be checked in tomorrow if the tinderbox gods are smiling on us. For now, EmailAddressAttributeEditor simply
calls SetValue() on the text control, which is the same thing that happens without a delegate at all.

AETypeOverTextCtrl.delegate was renamed to AETypeOverTextCtrl.staticControlDelegate; not only is it more descriptive,
but it fixes an apparent collision with a previously-existing 'delegate' attribute somewhere. The methods to get and
set the delegate (sorry, staticControlDelegate) were renamed similarly as well.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@12436 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe