Added support for -C/--continue to not stop during individual test runs if an error occurs. Currently unable to implement this for Suite runs because unittest.py doesn't support it.
Added support for -t/--test.  The name given to run is located within the list of all tests and, if found, added to a test list that is then run.
This allows for not only single test names (if unique) but also tests that share a common name in their parent list.

Refactored some of the code to remove duplication.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@12666 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe