YAATFTLTBF
(Yet another attempt to fix the linux tinderbox failures):

The logs show that we are trying to make an edited state change on an item
in the SMTPClient view; this leads to conflicts in some cases when that item
has been marked edited in the UI view.

It turns out this is happening because the TransportWrapper() class in
Application.py doesn't understand Annotations (or Stamps). So, when we try
to pass a MailStamp object to the UI view, we don't translate it into an
object in the correct view. So, I fixed this by making this class
Annotation-aware. 

git-svn-id: http://svn.osafoundation.org/chandler/trunk@12755 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe