- Part of Bug 7595, Add a dialog to turn on timezones when publishing/subscribing.
Now whenever timezones are turned on, existing unshared floating
events are converted to your current timezone.  Also, a prompt to
turn on timezones pops up when you start to publish a collection.

Still to do, make that dialog pop up in other timezone interaction
cases (when subscribing to a collection with timezones, or importing
timezoned events, or receiving an emailed event with a timezone).


git-svn-id: http://svn.osafoundation.org/chandler/trunk@12866 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe