Fix bug 6912 (Prioritize event and tickler dates over date send and last modified date in Date column), and bug 7763 (Triage status widget in Detail View needs tweaks).
For 6912: The addDisplayDates methods now take a 'now' parameter (both for perf and so that everyone uses the same idea of 'now'), since event dates in the future & past have different priorities. The priorities themselves are spaced adhoc integers - this might be improvable, but we don't persist them, so we can easily change them later... they're: 
10 Future event start time (if the item is an event, and its start time is in the future)
10 Future user reminder (if the item has a user reminder whose firing time is in the future)
20 Past event start time (if the item is an event whose start time is in the past)
30 Past user reminder (if the item has a user reminder whose firing time is in the past)
40 Date taken (photo only)
40 Timestamp of a user notification (user notification only)
50 Date sent (sent email only)
60 Date last modified

(Same priorities compete on ascending date value, of course).

For 7763: Resurrected original versions of the markupbar triage icons (my previous resurrection went haywire for some reason), and change the detail view to use them.

Also, fixes Reminder column icon display to always correspond to the Date column source: if it's a reminder time, the reminder icon will be shown; if it's an event start time, the little clock will be shown; otherwise, no icon.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@12883 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe