osaf.pim no longer has a dependency on osaf.sharing (almost); plus, first pass at edit/update API
- Added sharing.SharedItem stamp
- Moved sharing-related attributes and methods out of ContentItem and into sharing.SharedItem stamp; updated all code referring to these
- Removed collection's isReadOnly( )
- ContentItem.isAttributeModifiable( ) now delegates to any stamps an item might have, thereby severing the direct dependency on sharing layer

Now the only two places where osaf.pim has a dependency on sharing
are in calendar/Calendar.py and calendar/TimeZone.py.  The Occurrence
class has an iterModifiedAttributes( ) method which compares
occurrence values against the master.  I had to modify this to
ignore the SharedItem stamp on the master.  In TimeZone.py there
is a convertFloatingEvents( ) function which needs to examine the
shared state of an item, and the only way to do this is to import
osaf.sharing.

- Refactored merge( ) into its own module for use by sharing conduits and edit/update
- Fixed parsing of Cosmo "self" links; Chandler, given a "pim" URL can determine and GET the morsecode URL
- EIMML tweaks for Cosmo compatibility, and to allow same code to be used for Cosmo and edit/update



git-svn-id: http://svn.osafoundation.org/chandler/trunk@12963 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe