Add new "From" field to the email detail view. This manipulates a Chandler attribute called "originators", which is a list of EmailAddress objects. It has nothing to do with the Send As in the byline, which manipulates the Chandler attribute named "fromAddress", which maps to & from the email address's "From:" header.
There are issues about how this attribute should be set by default; after I do some work on the byline, I'll be in a better position to work on that. For now, it's empty, and doesn't interact with any other fields, but if you fill it in and send your message, the Chandler recipients you send it to should see whatever you put in it.

Also, the Who column doesn't yet consider this field - I'm not sure if it should.

git-svn-id: http://svn.osafoundation.org/chandler/trunk@13514 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe