Improvements to dump and reload:
We no longer dump and reload items in //parcels by default. This ensures we don't get new copies of items in //parcels and improves performance

Removed OTBCollectionMembershipRecord used by the sidebar and use the existing CollectionMembershipRecord instead. This allows the code to be simplified.

The sidebar now gets a fixed UUID so that reload can find it after a new repository is created.

Finish export exports CollectionMembershipRecords for the sidebar

Dump and Reload currently works, except for 2 workaround that pje is preparing fixes for.

git-svn-id: http://svn.osafoundation.org/chandler/trunk@13589 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe