Fix bug 8383 (Read-only subscribers should not be able to edit triage status if triage status is shared): the icon attribute editor (used for the triage column as well as the three icon-based columns) and the detail-view triage button now (indirectly) consider item.isAttributeModifiable.
This only partly fixes the underlying problem; not fixed yet: the nuances of whether a given read-only share should allow certain local changes to things anyway (like read/unread, or triagestatus if you're not sharing it). I'm going to close bug 8383 now because the rest of that is better spelled out in bug 8651 (users should be able to view / edit read-only items in the following way...), and it'll all be different when we switch to new sharing anyway (so it may just work).

git-svn-id: http://svn.osafoundation.org/chandler/trunk@13812 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe