Rearranged how exceptions are processed by the sharing layer
Exceptions can be "annotated" by upper layers to provide more context for
the error.  All the annotations are combined into a string that is stored
on the Share item for later viewing.

Also, an exception during pub/sub is displayed in a dialog immediately. We
can make this prettier later, but for now it's more important to get the right
debugging info.  There have been too many times where a bug report is missing
the critical piece of chandler.log.

Also, I'm activating monotlithic ICS syncing over EIM.  I will be removing
SimpleHTTPConduit as soon as the EVDB plugin can be updated.

Also, added a mergeFunction to recordset_conduit's commit( ) call



git-svn-id: http://svn.osafoundation.org/chandler/trunk@13888 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe