Guard against stale Parcel items while using schema.ns( )
How can this happen?

The osaf.framework.MasterPassword parcel does not have its installParcel()
called at startup, and so if a subscribe operation happens to be the first
thing that causes the MasterPassword parcel is installed, *and* that subscribe
fails -- thus causing the view to be cancelled -- the schema.ns( ) cache will
now contain a stale Parcel item.




git-svn-id: http://svn.osafoundation.org/chandler/trunk@13890 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe