EIM filtering improvements
- EIM state objects now do not include filtered-out values in the "agreed"
recordset so that when a filter is removed, inbound changes will, in fact,
be detected as changes.

- Because Cosmo assigns a default value of None for event status, and an empty
string for triage, I'm adding special converters to change those to NoChange
when creating records to avoid bogus conflicts.

- When a filter is removed by the user, we do a force-update sync which
refetches all items in the collection from the server.  Not doing so can
result in overwriting data.  Fixes bug 8793.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@13976 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe