Chandler will now try to clean up after a failed publish attempt
Now that we're breaking up a publish (or update) into multiple chunks it is
possible that the first chunk succeeds in creating the collection, but a later
chunk fails.  We want to remove the collection from the server so that a later
republish won't collide with a UUID already on the server.



git-svn-id: http://svn.osafoundation.org/chandler/trunk@14100 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe