Fix for bug 9108 (Sync receives deleted records, Chandler ignores)
What was happening was Chandler wasn't actually ignoring the deletes, it was
noticing that there were local changes to the remotely deleted item, and by
design it put the item back to the server.  The problem was that the sharing
layer wasn't being picky enough about what local changes should be considered,
and now it uses the EIM diff machinery to tell.

This commit also fixes a problem where changes committed by other views while
a sync was happening could be missed during the next sync.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@14277 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe