What started out as a quick fix for Bug 8883, keep mini-calendar highlighting consistent, Bug 7543, removing recurrence not refreshing mini-calendar, and Bug 8884, preview pane not updating when changing views, turned out to touch lots of things.
Investigating removing recurrence's impact on the mini-calendar led to
the discovery of a notification storm that should be fixed by Andi's
recent change to avoid propagating add notifications for items that
are deferring deletes plus the numeric index on nonRecurringNotes this
check in adds.

Also added code to ignore add notifications in the calendar/preview
area/minicalendar about dead items (I'm still seeing some of these
when removing recurrence).

To keep the preview area and mini-calendar in sync with the current
summary view, added a global CPIA event called ViewChanging that gets
propagated when a new tree of blocks is installed on the
SidebarBranchPointBlock.

Finally, tweaked the mini-calendar rendering logic to re-calculate busy
bars less often (but DO recalculate when recurrence rules change, which
wasn't happening before).



git-svn-id: http://svn.osafoundation.org/chandler/trunk@14286 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe