Bug 8791, make sure sharing icons get displayed.
This makes GetRawImage a little slower, but also makes it a little safer to use, because by default
it will now give a copy of the image it loaded from disk (or retrieved from cache) so that it is
safe to modify the image. More performant way would do it the way it was, where callers who were
going to modify the image made a copy themselves. As it is, we are doing some copying that is not
needed.

git-svn-id: http://svn.osafoundation.org/chandler/trunk@14478 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe