Backing out Andi's revision 14506 change to clear the current tinderbox test failures.
The reason for the revert is that while generate.py was
moved to the Chandler-debug plugin the following unit tests
still referenced osaf.pim.generate: TestCalendar, TestContacts and 
TestStamping.

I looked into adding the code that headless uses to initialize the
plugin environment but that was not a simple change and I think
would introduce more dependencies to the unit tests than needed.
Moving those tests to be Functional Tests is also an option but
not one that I'm ready to make late in the evening.

svn merge -r 14506:14505 . 



git-svn-id: http://svn.osafoundation.org/chandler/trunk@14510 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe