Fix Bug 9420: AttributeError: 'SSSidebarIconButton' object has no attribute 'buttonState'
SSSidebarButtons are Items that aren't referenced by things outside the
repository, so they may be deleted when the repository prunes.

However, some of our mouse-handling code assumed that these items have a
previously assigned 'buttonState', which is not a persisted attribute. So,
stash buttonStates in a global dict keyed by UUID, so that they'll still be
accessible if the button is deleted and reinstantiated by the repository.

git-svn-id: http://svn.osafoundation.org/chandler/trunk@14893 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe