Changes to Sharing Activity log (also fixes bug 8326 - Deleting sharing activity collection horks Chandler)
The "sharing activity" collection is no longer created out of the box, and
the Sync events are not created if the collection doesn't exist.  This will
avoid unneccesary log items.  You can create the sharing activity log via
the Tools > Sharing menu.  Deleting the collection will also prevent sync
events from being added later on.

The 'body' of the sync events is a bit easier to read now, as long as you speak EIM.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@15005 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe