- Fix bug 9499 and bug 9376 (which are roughly equivalent), when mailing an   occurrence of a recurring series, make the description match that   occurrence's details, not the master's.
Ideally we'd change the text to include details about the whole series, along
with the sent occurrence.  To do that requires changing localized strings, so
we'll that will need to be done using an English translation.  Added new 
arguments to be used in description localizations so those strings can
include details about the master.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@15008 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe