Fix for bug 10171, sharing issues stemming from the fact that a triage of LATER for a future recurring event is serialized the same as a DONE triage for an event in the past. r=morgen,grant
Also fixes assorted other minor problems with sharing
recurring events; the hasBeenSent field was
erroneously interpreted as True when an inbound
occurrence had hasBeenSent set to Inherit.

The change adds a new EIM field, lastPastOccurrence on
EventRecord.  This new field won't fix the problem for
collections shared in an 0.6.x Cosmo (i.e., the current
hub) because of Cosmo bug 10387, but it should be safe
to use against the old Cosmo.



git-svn-id: http://svn.osafoundation.org/chandler/trunk@15223 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe