Fix bug 10811 (Partial sync of events), hopefully for the last time, r=grant: my previous fix (r15419) slowed down a perf test, resulting in bug 11001.
To fix this (and the general case of a succession of programmatic attribute
changes made via a UI proxy), this patch uses a different callback: instead of
just committing, it calls a method on Application which will set a flag,
triggering a commit at the next idle. 

This reduces the number of extra substantive commits in the
PerfLargeDataTestStamping test from nine or so, to one.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@15431 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe