New PyLucene and chandlerdb from Andi
Quoth he:

Apart from the removal of the bug workaround in chandlerdb having to do with leaks of extension objects (such as the DbIndexOutput, DbIndexInput instances and so on in LuceneContainer.py) now that a proper fix is available, the PyLucene upgrade also contains an upgrade to Lucene 2.3.1.

One of the major improvements in Lucene 2.3 over earlier releases is indexing performance which was improved 3 to 4 fold.



git-svn-id: http://svn.osafoundation.org/chandler/trunk@16327 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe