Add makefile targets to skip plugin installation
With this addition, a "cd external; make world-no-plugins" does the
same thing as "make world", but skips the building of all plugins.

This is an ugly hack that should be considered temporary.  Building
locally, the gData plugin is erroring out because of not
finding the proper gdata.py upstream.  There's no explanation of why,
but adding these hacky Makefile targets allows me to build a full
Chandler for distribution without the plugin steps.  The plugins are
not needed for packaging anyway.

There are some cleaner ways to create these targets (making "world"
add "plugins" to the world-no-plugins target), but denormalizing them
keeps this hack from infecting the rest of the makefile.

Fixes and improvements welcome; I just need a way to avoid plugins for
now to advance on Ubuntu Hardy/Intrepid support.

See IRC transcript for more details:

http://chandlerproject.org/script/getIrcTranscript.cgi?channel=chandler&date=20081031&startTime=1450&endTime=1600


git-svn-id: http://svn.osafoundation.org/chandler/trunk@16598 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe