Update deb-building process to use libicu38 everywhere except gutsy
The mechanism for passing the proper libicu version (libicu36 or
libicu38) to the control file is pretty non-extensible/hacky but is a
reasonable way to use the existing build patterns.  If additional
variations are needed, a better system should probably be built.

The version numbers for the control file dependencies have been
dropped.  There's no real reason to require >=3.6 for libicu36 and
>=3.8 for libicu38; it works fine without it.


git-svn-id: http://svn.osafoundation.org/chandler/trunk@16600 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe